Makefile.am: Fix gnome-doc-utils.m4 inclusion
GNOME_DOC_DEFINES is used in configure.ac, thus aclocal needs to know
where to look for it.

It used to work by using the *installed* version of gnome-doc-utils.m4,
but for bootstrapping it fails. New features would be missing too.

While at it, add ${ACLOCAL_FLAGS} as adviced by the Automake manual.

https://bugzilla.gnome.org/show_bug.cgi?id=755966

Signed-off-by: Quentin Glidic <sardemff7+git@sardemff7.net>